+++ src/Presentation/Program.cs
@@ -1,7 +1,7 @@
 using Merrsoft.MerrMail.Application.Interfaces;
 using Merrsoft.MerrMail.Application.Services;
 using Merrsoft.MerrMail.Domain.Models;
-using Merrsoft.MerrMail.Infrastructure.Configuration;
+using Merrsoft.MerrMail.Infrastructure.Readers;
 using Merrsoft.MerrMail.Presentation;
 using Serilog;
 using Serilog.Events;
